# backend/src/lib/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FinAnalysis API",
    "version": "1.0.0",
    "description": "FinAnalysis API - Financial analysis and advisory platform. All endpoints are prefixed with `/api/v1`. Uses RFC 7807 Problem Details for errors."
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "Current API version"
    },
    {
      "url": "/api",
      "description": "Legacy API (deprecated)"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      },
      "csrfToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "access_token"
      }
    },
    "schemas": {
      "ProblemDetails": {
        "type": "object",
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "instance"
        ],
        "properties": {
          "type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string"
          }
        }
      },
      "ErrorResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ProblemDetails"
          }
        ]
      },
      "SuccessResponse": {
        "type": "object",
        "required": [
          "success",
          "data"
        ],
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "type": "object"
          }
        }
      },
...
    },
    "responses": {
      "ValidationError": {
        "description": "Validation error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "examples": {
              "validationError": {
                "summary": "Validation error",
                "value": {
                  "type": "https://tools.ietf.org/html/rfc7807#section-3.1",
                  "title": "Validation Error",
                  "status": 400,
                  "detail": "email: Invalid email format",
                  "instance": "/auth/login"
                }
              },
              "unauthorized": {
                "summary": "Unauthorized",
                "value": {
                  "type": "https://tools.ietf.org/html/rfc7807#section-3.1",
                  "title": "Unauthorized",
                  "status": 401,
                  "detail": "Invalid or expired token",
                  "instance": "/portfolio"
                }
              },
              "notFound": {
                "summary": "Not found",
                "value": {
                  "type": "https://tools.ietf.org/html/rfc7807#section-3.1",
                  "title": "Not Found",
                  "status": 404,
                  "detail": "Portfolio not found",
                  "instance": "/portfolio/550e8400-e29b-41d4-a716-446655440002"
                }
              }
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Unauthorized - Invalid or missing authentication",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "examples": {
              "validationError": {
                "summary": "Validation error",
                "value": {
                  "type": "https://tools.ietf.org/html/rfc7807#section-3.1",
                  "title": "Validation Error",
                  "status": 400,
                  "detail": "email: Invalid email format",
                  "instance": "/auth/login"
                }
              },
              "unauthorized": {
                "summary": "Unauthorized",
                "value": {
                  "type": "https://tools.ietf.org/html/rfc7807#section-3.1",
                  "title": "Unauthorized",
                  "status": 401,
                  "detail": "Invalid or expired token",
                  "instance": "/portfolio"
                }
              },
              "notFound": {
                "summary": "Not found",
                "value": {
                  "type": "https://tools.ietf.org/html/rfc7807#section-3.1",
                  "title": "Not Found",
                  "status": 404,
                  "detail": "Portfolio not found",
                  "instance": "/portfolio/550e8400-e29b-41d4-a716-446655440002"
                }
              }
            }
          }
        }
      },
...
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "examples": {
              "validationError": {
                "summary": "Validation error",
                "value": {
                  "type": "https://tools.ietf.org/html/rfc7807#section-3.1",
                  "title": "Validation Error",
                  "status": 400,
                  "detail": "email: Invalid email format",
                  "instance": "/auth/login"
                }
              },
              "unauthorized": {
                "summary": "Unauthorized",
                "value": {
                  "type": "https://tools.ietf.org/html/rfc7807#section-3.1",
                  "title": "Unauthorized",
                  "status": 401,
                  "detail": "Invalid or expired token",
                  "instance": "/portfolio"
                }
              },
              "notFound": {
                "summary": "Not found",
                "value": {
                  "type": "https://tools.ietf.org/html/rfc7807#section-3.1",
                  "title": "Not Found",
                  "status": 404,
                  "detail": "Portfolio not found",
                  "instance": "/portfolio/550e8400-e29b-41d4-a716-446655440002"
                }
              }
            }
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/portfolio": {
      "post": {
        "tags": [
          "portfolio"
        ],
        "summary": "Create portfolio",
        "operationId": "createPortfolio",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "assessmentId"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "assessmentId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "description": {
                    "type": "string"
                  }
                }
              },
              "examples": {
                "valid": {
                  "summary": "Create portfolio",
                  "value": {
                    "name": "Retirement Portfolio",
                    "assessmentId": "550e8400-e29b-41d4-a716-446655440001",
                    "description": "Long-term retirement savings"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Portfolio created"
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ]
      }
    },
    "/portfolio/{id}": {
      "get": {
        "tags": [
          "portfolio"
        ],
        "summary": "Get portfolio by ID",
        "operationId": "getPortfolio",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Portfolio found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                },
                "examples": {
                  "success": {
                    "summary": "Portfolio found",
                    "value": {
                      "success": true,
                      "data": {
                        "id": "550e8400-e29b-41d4-a716-446655440002",
                        "userId": "550e8400-e29b-41d4-a716-446655440000",
                        "assessmentId": "550e8400-e29b-41d4-a716-446655440001",
                        "name": "Retirement Portfolio",
                        "description": "Long-term retirement savings",
                        "totalValue": 1500000,
                        "createdAt": "2024-01-15T10:45:00.000Z",
                        "updatedAt": "2024-01-15T10:45:00.000Z"
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ]
      },
      "put": {
        "tags": [
          "portfolio"
        ],
        "summary": "Update portfolio",
        "operationId": "updatePortfolio",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Portfolio updated"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "portfolio"
        ],
        "summary": "Delete portfolio (soft delete)",
        "operationId": "deletePortfolio",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Portfolio deleted"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ]
      }
    },
...
    "/score/portfolio/{portfolioId}": {
      "get": {
        "tags": [
          "score"
        ],
        "summary": "Get scores for portfolio",
        "operationId": "getPortfolioScores",
        "parameters": [
          {
            "name": "portfolioId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Scores found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                },
                "examples": {
                  "success": {
                    "summary": "Portfolio scores",
                    "value": {
                      "success": true,
                      "data": [
                        {
                          "id": "550e8400-e29b-41d4-a716-446655440003",
                          "portfolioId": "550e8400-e29b-41d4-a716-446655440002",
                          "total": 85,
                          "goalAlignment": 90,
                          "assetAlloc": 80,
                          "diversification": 85,
                          "discipline": 88,
                          "efficiency": 82,
                          "tag": "OPTIMIZED",
                          "insights": "Well-diversified portfolio with good goal alignment.",
                          "createdAt": "2024-01-15T11:00:00.000Z",
                          "updatedAt": "2024-01-15T11:00:00.000Z"
                        }
                      ]
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ]
      }
    },
...
  }
}